+++ packages/pluggableWidgets/file-uploader-web/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- We fixed an issue where file uploader can still add more files when refreshed eventhough the number of maximum uploaded files has been reached.
+
+### Changed
+
+- We change the max file configuration to set maximum number of uploaded files through expression.
+
 ## [2.2.2] - 2025-07-01
 
 ### Fixed